+++ src/main/java/com/arcanc/biomorphosis/data/BioSpriteSourceProvider.java
@@ -13,6 +13,8 @@ import com.arcanc.biomorphosis.content.gui.BioSlot;
 import com.arcanc.biomorphosis.util.Database;
 import net.minecraft.client.renderer.texture.atlas.sources.SingleFile;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.particles.ParticleType;
+import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
@@ -24,6 +26,7 @@ import java.util.concurrent.CompletableFuture;
 public class BioSpriteSourceProvider extends SpriteSourceProvider
 {
     public static final ResourceLocation GUI_ATLAS = ResourceLocation.withDefaultNamespace("gui");
+    public static final ResourceLocation PARTICLE_ATLAS = ResourceLocation.withDefaultNamespace("particles");
 
     public BioSpriteSourceProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, ExistingFileHelper existingFileHelper)
     {
@@ -34,10 +37,21 @@ public class BioSpriteSourceProvider extends SpriteSourceProvider
     protected void gather()
     {
         //-------------------------------------------------------------------------------
+        // PARTICLES
+        //-------------------------------------------------------------------------------
+        atlas(PARTICLE_ATLAS).addSource(new SingleFile(
+                Database.rl("particle/hive_deco_fly"),
+                Optional.of(Database.rl("hive_deco"))));
+
+        //-------------------------------------------------------------------------------
         // TOOLTIPS
         //-------------------------------------------------------------------------------
-        atlas(GUI_ATLAS).addSource(new SingleFile(Database.rl("gui/tooltip/special_background"), Optional.of(Database.rl("tooltip/special_background"))));
-        atlas(GUI_ATLAS).addSource(new SingleFile(Database.rl("gui/tooltip/special_frame"), Optional.of(Database.rl("tooltip/special_frame"))));
+        atlas(GUI_ATLAS).addSource(new SingleFile(
+                Database.rl("gui/tooltip/special_background"),
+                Optional.of(Database.rl("tooltip/special_background"))));
+        atlas(GUI_ATLAS).addSource(new SingleFile(
+                Database.rl("gui/tooltip/special_frame"),
+                Optional.of(Database.rl("tooltip/special_frame"))));
 
         //-------------------------------------------------------------------------------
         // SLOTS