+++ src/main/java/com/arcanc/biomorphosis/util/Database.java
@@ -235,6 +235,7 @@ public class Database
 				public static final PageInfo V261 = new PageInfo(rl("v261"));
 	            public static final PageInfo V2611 = new PageInfo(rl("v2611"));
 	            public static final PageInfo V262 = new PageInfo(rl("v262"));
+	            public static final PageInfo V2632 = new PageInfo(rl("v2631"));
                 public static final PageInfo FLESH = new PageInfo(rl("flesh"));
                 public static final PageInfo NORPH_SOURCE = new PageInfo(rl("norph_source"));
                 public static final PageInfo CHAMBER = new PageInfo(rl("chamber"));